siglap: Add ARM libraries
In bad form, these ARM libraries are binary blobs.

They come from the official ARM versions of GCC, and are built for the
ARMv6-M chips.

Most of the development for Siglap is done on Novena, where the native C
compiler can generate cortex-m0plus binaries.  However, the support
libraries are designed for the ARMv7-A, and have instructions that are
not valid on the older parts.

This commit adds binary libraries for libgcc -- which includes things like
division support and modular arithmetic -- and binaries for libc-nano --
which gives basic memcpy/strcpy/etc. support.

Since these are redistributed from ARM, please make any enqueries to them
regarding the source of these libraries.

Signed-off-by: Sean Cross <xobs@kosagi.com>