refactor LED interface to include UI mechanism
The LED drivers -- for the 16 UI LEDs on the front -- now has
the concept of separate UI and effects frame buffers.

The effects animations render to an effects frame buffer. UI
cues render to a UI buffer. These are merged at every effects
update increment and then sent over to to the LED array via
the bitbang loop.

The merge is a simple saturating add of the colors.

This allows for the jog dial to now have an effect associated
with it which is overlaid onto the ambient color effect.