gpiox: Fix INT line reading
The INT line is apparently the AND of the eight IRQ lines PLUS the status
of the reset operation.  That is, when the SW_RST operation completes, a
bit (1 << 1) is set in the ID register (0x01) that indicates reset is done,
and the INT/ line is driven low.

In order to make use of the interrupt line, we must read the ID register
until RST_INT goes high, which will clear the bit and unstick the IRQ/ line.

Signed-off-by: Sean Cross <xobs@kosagi.com>