fix responsivity bug in transitions between charger and boost modes
a bit of a hack -- when we call boost or charger to turn off, we immediately
clear the watchdog and boost bits.

This is done because we can't clear the watchdog and boost bits in the thread
itself; for some reason it seems to act as a watchdog timer wakeup.