Add Flash memory support
This is a port in of the C90TFS reference code from Freescale (which has
a permissive license allowing open source so long as their copyright notice
and disclaimers are preserved, and we don't use the Freescale name in vain).

Most of the code is in ext/C90TFS. We bodge the machine ID by faking us as
an MKL25Z4, which is our closest relative that's supported in their BSP. From
the FTL standpoint, we're identical.

A set of wrappers around their awful code is provided in flash.c. Currently,
we support erasing and programming, as well as security sector read out.
Features that might be nice to add are things like setting the security
sectors, manipulating the IFR (basically read-only parameters with a write-
once option for cal data), and the protection fields.

A few notes about potential bodges:

- Erasing is specified by sector, but programming is specified by actual address.
I sort of like this inconsistency because it reminds the user that you're going
to erase a whole lot of data, whereas programming can happen on a word-by-word
basis.

- Flash programming data must be sourced from RAM. I check the bounds of this
using hard-coded RAM addresses. It works for this device but it's not portable.
I presume there is some clever chibiOS macro to do this but I don't know what it is.

You know what, as I write that above paragraph and I'm looking at the code,
I'm lying about not being able to program Flash to Flash. I will fix this in
a future commit.