Remove check for program data source in RAM
After reading the source code for the C90TFS implementation, the
word-by-word programming explicitly copies data to an FCCOB register
prior to kicking off the program command. As such, it should be safe
to read data from one location in flash before programming to another.

Due to the erasure checks built into the program routine, you shouldn't
be able to overwrite your own code data without the system crashing.

The code /is/ structured to allow self-modification of the code due
to the use of RAM callbacks, but this is tricky and for the sake of
simplicity we'll just not allow it. If we have to support OTA updates
down the road, this issue needs to be revisited and tested thoroughly.