prevent home button sequence from passing on to apps
Also, add a debounce interval that's larger for single button
hits than for jog dial hits, to give some margin for pressing the
buttons at the same time. If it's too hard to hit the buttons at
the same time, the margin can be increased for single button
hits without increasing lag on the jog dial.