clear some of the names up in the collection loop and tweak values
Unclear to me if the delay is working 100% as intended, but it does seem
that the menu mode is more reliable to exit without tripping up the existing
setting on the app below.

I tried increasing the collect timeout to 500ms, which should really
exaggerate the collection delay. There is a big delay on the UI, but
multiple events seem to trip up the timer system...not sure what this
means, the timer is supposed to reset and begin again but the system
seems to be working well enough and there are bigger bugs to smash.