Add ADC support.
ADC support enabled in ChibiOS. This commit also includes a
demo app that displays the temperature of the CPU.

The ADC HAL in general is structured so that you first call
an update() function to queue a sampling request from the ADC,
and then an event is issued which the application should handle
later on when the ADC completes.

Also, in this commit a small refactor of the board revision
system was brought in where the board rev (EVT1 vs. EVT1b etc.)
is now specified in the Makefile and not in the board.h file which
is in the larger chibiOS HAL tree.