change poweroff to standby behavior (not shipmode)
Standby is friendlier than ship mode. Ship mode will need a UI
escape to enter as well, but we'll design that app down the road.
Relatively low priority and ship mode could be something that is
entered via the factory test sequence only.

Add a UI message about how to power the system back on again.