add test framework to the code base
For now, it's just a dummy set of callbacks that report "no test
implemented".

But now, every low level driver has the skeleton hanging there to
install a test into the test framework.

A test audit log framework is also implemented. The audit log doesn't
do anything currently, but the callbacks are provided to record
how many times a test was run, and the last result of each test.