basic test audit log infrastructure complete
Audit logs update when tests are run with the type that is not
"power on". You could audit power-on tests but you run the risk of
wearing out flash, as the device only has a 10k erase cycle lifetime.

Currently, all tests have the status "not existing", so you get
a message at boot reminding you to write the test.