add interactive test modes to oled and led functions
Also added some helper functions for interactive tests, as well as
an app mode for interactive testing.

If an app is running, it could interfere with a test. So cmd-test
kills the current app running and launches the testmode app, which
is a dummy app that safely does nothing while the tests are being
run.