add radio test
Requires two orchard devices, a peer, and the DUT.

The peer can enter test mode by selecting the ~testpeer app.

The peer listens for packets on one port, and rebroadcasts them on
another port.

The DUT transmits its device ID plus a random number and checks that
the peer has rebroadcasted the packet.

The DUT will retry transmitting until a timeout interval is met,
after which it declares the test has failed.

The test is a little bit unreliable due to the peer crashing
mysteriously. It seems to be linked to a problem with receiving
and transmitting a packet at the same time?