add microphone and charger tests, and auditCheck() function
microphone test just runs the oscope app, and requires it to be
compiled in to succeed. Operator must answer a prompt for the
test to pass.

charger test covers charger, gas gauge, and USB subsystems. Checks
that the charger can be automatically detected and that the battery
and system push at least a minimum amount of current in both
charge and discharge states. The limits are pretty generous at
this point as we're just looking for gross defects.

auditCheck() function works, but it seems that calling it from
gdb crashes the system. My guess is that the ORFS storage layer
is too heavy to call out-of-band, but we'll have to look into this
later.