radio: Clear FIFO when CRC fails
When the CRC fails, it can leave data in the FIFO.  In this state, the IRQ
will be low, but since the FIFO is not empty, no new packets will arrive.

A workaround is to drain the FIFO manually, but this is suboptimal.

This patch enables CRC autoclear, which drains the FIFO when the CRC
is invalid.

Signed-off-by: Sean Cross <xobs@kosagi.com>