add the beginning of the radio ping protocol + rand() entropy freshening
Part of the yak shaving to the radio ping protocol was improving rand().
Had to pull in some partial changes on the radio ping to also check in
the final improvements to rand().

I don't recommend you use rand() for cryptographic numbers but as
a source of randomness for effects and timing it's probably okay.