hack to create & manage a family of lightgenes
This is pretty awful, but hopefully doesn't need much modification.

Any effect that begins with the characters "lg" is now a lightgene
effect. The character immediately after lg refers to the individual
index, e.g. lg2 refers to index 2 of the individual genome array.

Whenever a new effect index is selected, we check to see if it's
a lightgene, and if so, compute the diploid genome based upon
the haploid halves.

We also check when doing the UI rendering in app-default for the
lg case and replace the lg effect name with the name of the diploid
individual.