add userconfig block and UI
userconfig is a block of data for storing user configuration options.

These are small bits of data like performance counters that can change
and need to be remembered between resets.

Curretly commits are "lazy" e.g. only happen if the user actually
changes anything and the poll happens concommitant with the USB
charger detection routine.